+++ Robust.UnitTesting/Shared/Localization/LocalizationTests.cs
@@ -4,7 +4,6 @@ using System.Globalization;
 using NUnit.Framework;
 using Robust.Shared.ContentPack;
 using Robust.Shared.GameObjects;
-using Robust.Shared.GameObjects.Components.Localization;
 using Robust.Shared.IoC;
 using Robust.Shared.Localization;
 using Robust.Shared.Prototypes;